requirements.txt
setuptools~=68.2.2
yaspin~=3.0.1
feedparser~=6.0.10
beautifulsoup4
titlecase
PyYAML
gensim
scikit-learn
